refactor: move post info out of main package
