--- packages/maskbook/src/plugin-infra/utils/index.tsx
@@ -1,51 +1,0 @@
-import type { Plugin } from '../types'
-import { useEffect, useState, useRef } from 'react'
-import { ErrorBoundary } from '@dimensiondev/maskbook-theme'
-type Inject<T> = Plugin.InjectUI<T>
-type Raw<T> = Plugin.InjectUIRaw<T>
-
-export function createInjectHooksRenderer<PluginDefinition extends Plugin.Shared.Definition, PropsType>(
-    usePlugins: () => PluginDefinition[],
-    pickInjector: (plugin: PluginDefinition) => undefined | Inject<PropsType>,
-) {
-    const picker = (plugin: PluginDefinition) => ({
-        key: plugin.ID,
-        name: plugin.name,
-        ui: pickInjector(plugin),
-    })
-    return function InjectHooksRenderer(props: PropsType) {
-        const all = usePlugins()
-            .map(picker)
-            .filter((x) => x.ui)
-            .map(({ key, name, ui }) => (
-                <ErrorBoundary key={key} subject={`Plugin ` + name.fallback}>
-                    <Main UI={ui!} data={props} />
-                </ErrorBoundary>
-            ))
-        return <>{all}</>
-    }
-}
-
-function Main<T>(props: { data: T; UI: Inject<any> }) {
-    const { data, UI } = props
-    if (isRawInjectHook(UI)) return <RawHookRender UI={UI} data={data} />
-    return <UI {...data} />
-}
-function RawHookRender<T>({ UI, data }: { data: T; UI: Raw<T> }) {
-    const [ref, setRef] = useState<HTMLDivElement | null>()
-    const [f, setF] = useState<(props: T) => void>()
-    const cancel = useRef<AbortController>()
-
-    useEffect(() => {
-        if (!ref) return
-        const sig = (cancel.current = new AbortController())
-        setF(UI.init(sig.signal, ref))
-        return () => sig.abort()
-    }, [ref, UI.init])
-    useEffect(() => void f?.(data), [f, data])
-
-    return <div ref={(r) => ref === r || setRef(r)} />
-}
-function isRawInjectHook<T>(x: Inject<T>): x is Raw<T> {
-    return 'type' in x && x.type === 'raw'
-}
