+++ packages/maskbook/tsconfig.json
@@ -8,6 +8,11 @@
   // ! Note: maskbook/ depends dashboard/ on source-code level
   // !       but dashboard/ depends maskbook/ on type level
   // !       do not add dashboard/ as a reference otherwise there is a circular dependency on the type level.
-  "references": [{ "path": "../shared/" }, { "path": "../theme/" }, { "path": "../icons/" }],
+  "references": [
+    { "path": "../shared/" },
+    { "path": "../theme/" },
+    { "path": "../icons/" },
+    { "path": "../plugin-infra" }
+  ],
   "ts-node": { "transpileOnly": true, "compilerOptions": { "module": "CommonJS" } }
 }
