+++ packages/plugin-infra/package.json
@@ -3,9 +3,12 @@
   "private": true,
   "license": "AGPL-3.0-or-later",
   "scripts": {},
+  "types": "./dist/",
   "dependencies": {
-    "@servie/events": "^3.0.0",
+    "@dimensiondev/holoflows-kit": "0.8.0-20210317064617-6c4792c",
     "@dimensiondev/maskbook-shared": "workspace",
-    "@dimensiondev/maskbook-theme": "workspace"
+    "@dimensiondev/maskbook-theme": "workspace",
+    "@servie/events": "^3.0.0",
+    "ts-results": "^3.2.1"
   }
 }
