+++ packages/plugin-infra/src/index.ts
@@ -3,3 +3,4 @@ export * from './manager/dashboard'
 export * from './manager/sns-adaptor'
 export * from './manager/worker'
 export * from './utils'
+export * from './PostInfo'
