+++ packages/shared/package.json
@@ -3,9 +3,11 @@
   "dependencies": {
     "@dimensiondev/holoflows-kit": "0.8.0-20210317064617-6c4792c",
     "@servie/events": "^3.0.0",
+    "anchorme": "^2.1.2",
     "async-call-rpc": "^5.0.0",
     "bignumber.js": "^9.0.1",
     "immer": "^9.0.1",
+    "lodash-es": "^4.17.21",
     "ts-results": "^3.2.0",
     "typeson": "^6.1.0",
     "typeson-registry": "^1.0.0-alpha.38",
