+++ packages/shared/src/index.ts
@@ -9,3 +9,4 @@ export * from './react/createGlobalState'
 export * from './ShadowRoot'
 export * from './TypedMessage'
 export * from './web3'
+export * from './utils'
