+++ packages/shared/src/utils/parseURL.ts
@@ -1,0 +1,4 @@
+import parser from 'anchorme'
+export function parseURL(string: string) {
+    return parser.list(string).map((x) => x.string)
+}
